+++ README.md
@@ -1,6 +1,10 @@
 # RPG
-## Description
 Interactive RPG game about being in a mansion... And that's it, I guess?
+
+## Requirements
+* python :: >= 3.0
+* make :: (optional)
+
 ## Install
 ```bash
 git clone https://Eyv-cyber/cool-rpg ~/cool-rpg
